Conflicts:
	pychron/experiment/automated_run/automated_run.py
	pychron/spectrometer/thermo/detector.py

+++ pychron/updater/packager.py
@@ -31,7 +31,8 @@ def copy_resources(root, dest, app_name):
 
     # copy icons
     iroot = os.path.join(root, 'resources', 'icons')
-    idest = os.path.join(dest, 'icons')
+    rdest = os.path.join(dest, 'Resources')
+    idest = os.path.join(rdest, 'icons')
     if not os.path.isdir(idest):
         os.mkdir(idest)
 
@@ -56,14 +57,14 @@ def copy_resources(root, dest, app_name):
     # copy helper mod
     for a in ('helpers', ):
         m = os.path.join(root, 'launchers', '{}.py'.format(a))
-        copy_resource(dest, m)
+        copy_resource(rdest, m)
 
     # copy qt_menu.nib
     p = '/anaconda/python.app/Contents/Resources/qt_menu.nib'
     if not os.path.isdir(p):
         p = '{}/{}'.format(os.path.expanduser('~'),
                            'anaconda/python.app/Contents/Resources/qt_menu.nib')
-    copy_resource_dir(p)
+    copy_resource_dir(rdest, p)
 
 
 def make_egg(root, dest, pkg_name, version):
@@ -77,39 +78,45 @@ def make_egg(root, dest, pkg_name, version):
     try:
         setup(name=pkg_name,
               script_args=('bdist_egg',),
-              version=version,
               packages=pkgs)
     except BaseException, e:
         import traceback
         traceback.print_exc()
 
-    eggname = '{}-{}-py2.7.egg'.format(pkg_name, version)
+    eggname = '{}-0.0.0-py2.7.egg'.format(pkg_name)
     # make the .pth file
+
     if dest.endswith('Contents'):
-        with open(os.path.join(dest,
-                               'Resources',
+        rdest = os.path.join(dest, 'Resources')
+        with open(os.path.join(rdest,
                                '{}.pth'.format(pkg_name)), 'w') as fp:
             fp.write('{}\n'.format(eggname))
 
         egg_root = os.path.join(root, 'dist', eggname)
-        copy_resource(dest, egg_root)
+        copy_resource(rdest, egg_root)
 
     # remove build dir
     for di in ('build', 'dist','pychron.egg-info'):
         p = os.path.join(root, di)
         print 'removing entire {} dir {}'.format(di, p)
-        shutil.rmtree(p)
+        if os.path.isdir(p):
+            shutil.rmtree(p)
+        else:
+            print 'not a directory {}'.format(p)
 
 
 def resource_path(dest, name):
-    return os.path.join(dest, 'Resources', name)
+    return os.path.join(dest, name)
 
 
-def copy_resource_dir(src, name=None):
+def copy_resource_dir(dest, src, name=None):
     if os.path.exists(src):
         if name is None:
             name = os.path.basename(src)
-        shutil.copytree(src, resource_path(name))
+
+        rd = resource_path(dest, name)
+        if not os.path.exists(rd):
+            shutil.copytree(src, rd)
     else:
         print '++++++++++++++++++++++ Not a valid Resource {} +++++++++++++++++++++++'.format(src)
